feat: Add RandomMapKey and RandomMapValuet
These two functions are analog to RandomStringMapKey and
RandomStringMapValue, in the same way RandomElement is to
RandomStringElement.

This was something I felt was missing as I was trying to get a random
map element from a map than was not a map[string]string.

Unlike their String counterpart though, no sorting is performed. Doing a
sort would require the generics to be constrained to cmd.Ordered which I
feel is too restrictive.